# Conflicts:
#	go-plugin/configs/codecs/springcloud/egress_springcloud.json
#	go-plugin/configs/codecs/springcloud/ingress_springcloud.json
#	go-plugin/configs/codecs/springcloud/metadata.json
#	go-plugin/etc/ant/env_conf
#	go-plugin/go.mod
#	go-plugin/go.sum
#	go-plugin/vendor/github.com/klauspost/compress/LICENSE
#	go-plugin/vendor/github.com/klauspost/compress/flate/inflate.go
#	go-plugin/vendor/github.com/klauspost/compress/flate/regmask_other.go
#	go-plugin/vendor/github.com/klauspost/compress/gzip/gunzip.go
#	go-plugin/vendor/github.com/klauspost/compress/zlib/reader.go
#	go-plugin/vendor/github.com/rcrowley/go-metrics/.travis.yml
#	go-plugin/vendor/github.com/valyala/fasthttp/.gitignore
#	go-plugin/vendor/github.com/valyala/fasthttp/README.md
#	go-plugin/vendor/github.com/valyala/fasthttp/args.go
#	go-plugin/vendor/github.com/valyala/fasthttp/bytesconv.go
#	go-plugin/vendor/github.com/valyala/fasthttp/bytesconv_32.go
#	go-plugin/vendor/github.com/valyala/fasthttp/bytesconv_64.go
#	go-plugin/vendor/github.com/valyala/fasthttp/client.go
#	go-plugin/vendor/github.com/valyala/fasthttp/cookie.go
#	go-plugin/vendor/github.com/valyala/fasthttp/go.mod
#	go-plugin/vendor/github.com/valyala/fasthttp/go.sum
#	go-plugin/vendor/github.com/valyala/fasthttp/header.go
#	go-plugin/vendor/github.com/valyala/fasthttp/http.go
#	go-plugin/vendor/github.com/valyala/fasthttp/server.go
#	go-plugin/vendor/github.com/valyala/fasthttp/status.go
#	go-plugin/vendor/github.com/valyala/fasthttp/strings.go
#	go-plugin/vendor/github.com/valyala/fasthttp/tcpdialer.go
#	go-plugin/vendor/github.com/valyala/fasthttp/uri.go
#	go-plugin/vendor/github.com/valyala/fasthttp/uri_unix.go
#	go-plugin/vendor/github.com/valyala/fasthttp/uri_windows.go
#	go-plugin/vendor/github.com/valyala/fasthttp/userdata.go
#	go-plugin/vendor/modules.txt
#	go-plugin/vendor/mosn.io/api/network.go
#	go-plugin/vendor/mosn.io/api/route.go
#	go-plugin/vendor/mosn.io/api/stream_filter.go
#	go-plugin/vendor/mosn.io/api/xprotocol.go

+++ go-plugin/etc/script/compile.sh
@@ -25,4 +25,8 @@ CGO_ENABLED=1 go build -mod=readonly -gcflags "all=-N -l" \
   -ldflags "-B 0x$(head -c20 /dev/urandom | od -An -tx1 | tr -d ' \n') -X main.Version=${MAJOR_VERSION} -X main.GitVersion=${GIT_VERSION}" \
   -v -o mosn "${SIDECAR_PROJECT}/cmd/mosn/main"
 
-mv mosn "/go/src/${PLUGIN_PROJECT}/build/sidecar/binary/mosn"
+if [ -f mosn ]; then
+  md5sum -b mosn | cut -d' ' -f1 >mosn-${MAJOR_VERSION}-${GIT_VERSION}.md5
+  mv mosn-${MAJOR_VERSION}-${GIT_VERSION}.md5 "/go/src/${PLUGIN_PROJECT}/build/sidecar/binary/mosn-${MAJOR_VERSION}-${GIT_VERSION}.md5"
+  mv mosn "/go/src/${PLUGIN_PROJECT}/build/sidecar/binary/mosn"
+fi
